Updates and Bug Fixes
There was a TON of cleaning to do. I finished properly merging the code and fixed the bugs from that. Then I realized that I had made a mistake in the design of the program, so I went and edited/created a few methods to solve the problem(s) and fixed bugs related to my changes. I also made the chicken start on even numbered rows and columns, because otherwise the game is literally impossible to win.